Removing average from results
